Simple Login Page UI
I added a LoginControl component, along with a few subfunctions to show
a visible login page with a username and password field, log-in button,
register button, and logout button. I put in some code for handling
logging in and registering for the sake of visible functionality (right
now it just works with an array of users that are hard coded in there),
but we can delete and replace that with database functionality once
that's up and running. Also, I've put the isLoggedIn and loggedInUser
states in the loginControl for now, but was wondering if they'll need to
be pulled up all the way to App since the whole page will need to know
whether you're logged in for it to decide what to display. Just some
thoughts.