Revisions to tests and added login test
The login test does not work yet. I've revised the database client tests
to how they are supposed to work with callbacks, but....the tests all
fail now. It seems that all the calls to client time out before
finishing, probably because the server isn't running. How do we run the
tests without the server???