More test updates
I updated the LogInForm test so that it can correctly test if it calls
the getUser function that is passed into it. I think this is why the
react apps I have seen have components take in methods rather than
import.

ALSO, to do this test, I installed enzyme and react-addons-test-utils.
If you want it to work you should run npm install again. I also had to
add a new class to the CSS file