Had to fix getTotalPBIPercentage
It wasn't quite working right. Was returning 0 every time. Also, since
for the sake of creating tasks we need to know the total percentage of
tasks CREATED (not just in the done column), I made a new database
function for that instead. Also, changed editTaskForm so that it keeps
track of the old value of the percentage and subtracts that from the
total. Before, the tasks old percentage counted against a total, so, for
instance, if a task was at 80% and you were trying to reduce it to 30%,
it wouldn't let you since the current "total" is 80% which, when added
to 30 is over 100. But it works now